+++ Source/SwiftLintBuiltInRules/Rules/Idiomatic/ReduceIntoInsteadOfLoop.swift
@@ -1,5 +1,8 @@
 import SwiftSyntax
 
+typealias ReferencedVariable = ReduceIntoInsteadOfLoop.ReferencedVariable
+typealias CollectionType = ReduceIntoInsteadOfLoop.CollectionType
+
 struct ReduceIntoInsteadOfLoop: ConfigurationProviderRule, SwiftSyntaxRule, OptInRule {
     var configuration = SeverityConfiguration<Self>(.warning)
 
@@ -19,7 +22,7 @@ struct ReduceIntoInsteadOfLoop: ConfigurationProviderRule, SwiftSyntaxRule, OptI
     }
 }
 
-fileprivate extension ReduceIntoInsteadOfLoop {
+internal extension ReduceIntoInsteadOfLoop {
     final class Visitor: ViolationsSyntaxVisitor {
         override func visitPost(_ node: CodeBlockItemListSyntax) {
             // reduce into forInStmts and variableDecls map
@@ -387,52 +390,3 @@ private extension VariableDeclSyntax {
     }
 }
 
-private struct ReferencedVariable: Hashable {
-    let name: String
-    let position: AbsolutePosition
-    let kind: Kind
-
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(self.name)
-        hasher.combine(self.position.utf8Offset)
-        hasher.combine(self.kind)
-    }
-
-    enum Kind: Hashable {
-        case method(name: String, arguments: Int)
-        case assignment(subscript: Bool)
-
-        func hash(into hasher: inout Hasher) {
-            switch self {
-            case let .method(name, arguments):
-                hasher.combine("method")
-                hasher.combine(name)
-                hasher.combine(arguments)
-            case let .assignment(`subscript`):
-                hasher.combine("assignment")
-                hasher.combine(`subscript`)
-            }
-        }
-    }
-}
-
-private struct CollectionType: Hashable {
-    let name: String
-    let genericArguments: Int
-
-    static let types: [CollectionType] = [
-        .set,
-        .array,
-        .dictionary
-    ]
-
-    static let array = CollectionType(name: "Array", genericArguments: 1)
-    static let set = CollectionType(name: "Set", genericArguments: 1)
-    static let dictionary = CollectionType(name: "Dictionary", genericArguments: 2)
-
-    static let names: [String: CollectionType] = {
-        return CollectionType.types.reduce(into: [String: CollectionType]()) { partialResult, collectionType in
-            partialResult[collectionType.name] = collectionType
-        }
-    }()
-}
